Render GitHub flavoured markdown (#41)
We're getting readable output on GHA runs, but I think it should
actually render into HTML and it's not. Pretty sure this is because
`py_markdown_table` renders tables with `+` at the corner, but I think
'Github flavoured' Markdown wants `|` all up and down the columns.

This tries rendering a table by hand, which is a little less nice in
that it doesn't balance the column widths, but if it renders as it
should we won't even be reading the raw markdown anyway.

Also moves the code into the main lib and adds a smoke test (mostly just
to make sure that e.g. changes to the eval schema still render OK, since
that's sensitive to the contained types)