Add referrenced files to context (#43)
There's now a strict output schema for the files rather us just nicely
asking the LLM to produce a string in the right format.

We use this to parse out the snippets (with the 5 lines before and
after) and put these in the issues when presenting it to the eval agent.
Not made a huge difference in the estimates I've seen, but it does seem
a lot more likely to identify things as "local". I think this is mostly
an improvement.

Fixes #27